+++ middlewares/profile.mw.js
@@ -1,0 +1,5 @@
+const AppError = require("../utils/AppError")
+
+exports.restrictTo = (...profiles) => {
+    //TO DO
+}
